dev.requirements.txt
fastapi
httpx
maap_py
pydantic
pytest
python-dotenv
requests
uvicorn